Remove previous rich text editor
