+++ editor.planx.uk/package.json
@@ -6,10 +6,6 @@
     "@airbrake/browser": "^2.1.7",
     "@apollo/client": "^3.6.9",
     "@ctrl/tinycolor": "^3.4.1",
-    "@draft-js-plugins/anchor": "^4.1.4",
-    "@draft-js-plugins/buttons": "^4.3.1",
-    "@draft-js-plugins/editor": "^4.1.3",
-    "@draft-js-plugins/inline-toolbar": "^4.1.6",
     "@emotion/react": "^11.10.4",
     "@emotion/styled": "^11.10.4",
     "@feedback-fish/react": "^1.2.1",
@@ -45,8 +41,6 @@
     "classnames": "^2.3.1",
     "core-js": "^3.24.1",
     "date-fns": "^2.29.1",
-    "draft-js": "^0.11.7",
-    "draftjs-to-html": "^0.9.1",
     "fast-xml-parser": "^4.0.11",
     "formik": "^2.2.9",
     "graphql": "^16.5.0",
