+++ editor.planx.uk/src/ui/RichTextInput.stories.tsx
@@ -1,9 +1,9 @@
-import Box from "@mui/material/Box";
-import Button from "@mui/material/Button";
+import Box from "@material-ui/core/Box";
+import Button from "@material-ui/core/Button";
 import { Meta } from "@storybook/react/types-6-0";
 import React, { useState } from "react";
 
-import RichTextInput from "./RichTextInput";
+import RichTextInput, { fromHtml, injectVariables } from "./RichTextInput";
 
 const metadata: Meta = {
   title: "Design System/Atoms/Form Elements/RichTextInput",
@@ -22,7 +22,9 @@ export const Basic = () => {
     <Box marginTop={3}>
       <Button
         onClick={() => {
-          setValue("I am reset");
+          setValue(
+            `<p>Hello, and <span data-type="mention" class="pass" data-id="Apples">@Apples</span> <a target="_blank" rel="noopener noreferrer nofollow" href="https://opensystemslab.org">OSL</a></p>`
+          );
         }}
       >
         Reset from the outside
@@ -36,7 +38,28 @@ export const Basic = () => {
       />
       <Box>
         <p>HTML result:</p>
-        <pre>{value}</pre>
+        <textarea
+          style={{ display: "block", width: "100%", height: 120 }}
+          value={value}
+          readOnly
+        />
+      </Box>
+      <Box>
+        <p>Interpolated HTML result:</p>
+        <textarea
+          style={{ display: "block", width: "100%", height: 120 }}
+          value={injectVariables(value, { name: "Gary" })}
+          readOnly
+        />
+      </Box>
+
+      <Box>
+        <p>JSON result:</p>
+        <textarea
+          style={{ display: "block", width: "100%", height: 120 }}
+          value={JSON.stringify(fromHtml(value), null, 2)}
+          readOnly
+        />
       </Box>
     </Box>
   );
